+++ src/main/java/com/trainstation/service/GaService.java
@@ -2,6 +2,11 @@ package com.trainstation.service;
 
 import com.trainstation.dao.GaDAO;
 import com.trainstation.model.Ga;
+import com.trainstation.network.AppClient;
+import com.trainstation.network.AppRequest;
+import com.trainstation.network.AppResponse;
+import com.trainstation.network.NetworkConfig;
+import com.trainstation.network.RequestType;
 import java.util.List;
 
 /**
@@ -23,10 +28,20 @@ public class GaService {
     }
 
     public List<Ga> layTatCaGa() {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.GET_ALL_GA));
+            if (resp.isSuccess() && resp.getData() != null) return (List<Ga>) resp.getData();
+            return java.util.Collections.emptyList();
+        }
         return gaDAO.getAll();
     }
 
     public Ga timGaTheoMa(String maGa) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.FIND_GA_BY_ID, maGa));
+            if (resp.isSuccess()) return (Ga) resp.getData();
+            return null;
+        }
         return gaDAO.findById(maGa);
     }
 
@@ -50,14 +65,47 @@ public class GaService {
     }
 
     public boolean themGa(Ga ga) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.INSERT_GA, ga));
+            if (resp.isSuccess() && resp.getData() != null) return (Boolean) resp.getData();
+            return false;
+        }
         return gaDAO.insert(ga);
     }
 
     public boolean capNhatGa(Ga ga) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.UPDATE_GA, ga));
+            if (resp.isSuccess() && resp.getData() != null) return (Boolean) resp.getData();
+            return false;
+        }
         return gaDAO.update(ga);
     }
 
     public boolean xoaGa(String maGa) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.DELETE_GA, maGa));
+            if (resp.isSuccess() && resp.getData() != null) return (Boolean) resp.getData();
+            return false;
+        }
         return gaDAO.delete(maGa);
+    }
+
+    public List<Ga> layGaDaXoa() {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.GET_DELETED_GA));
+            if (resp.isSuccess() && resp.getData() != null) return (List<Ga>) resp.getData();
+            return java.util.Collections.emptyList();
+        }
+        return gaDAO.getDeletedStations();
+    }
+
+    public boolean khoiPhucGa(String maGa) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.RESTORE_GA, maGa));
+            if (resp.isSuccess() && resp.getData() != null) return (Boolean) resp.getData();
+            return false;
+        }
+        return gaDAO.restoreStation(maGa);
     }
 }
