+++ src/main/java/com/trainstation/service/NhanVienService.java
@@ -2,6 +2,11 @@ package com.trainstation.service;
 
 import com.trainstation.dao.NhanVienDAO;
 import com.trainstation.model.NhanVien;
+import com.trainstation.network.AppClient;
+import com.trainstation.network.AppRequest;
+import com.trainstation.network.AppResponse;
+import com.trainstation.network.NetworkConfig;
+import com.trainstation.network.RequestType;
 import java.util.List;
 
 /**
@@ -26,6 +31,11 @@ public class NhanVienService {
      * Lấy tất cả nhân viên
      */
     public List<NhanVien> layTatCaNhanVien() {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.GET_ALL_NHANVIEN));
+            if (resp.isSuccess() && resp.getData() != null) return (List<NhanVien>) resp.getData();
+            return java.util.Collections.emptyList();
+        }
         return nhanVienDAO.getAll();
     }
 
@@ -33,6 +43,11 @@ public class NhanVienService {
      * Tìm nhân viên theo mã
      */
     public NhanVien timNhanVienTheoMa(String maNV) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.FIND_NHANVIEN_BY_ID, maNV));
+            if (resp.isSuccess()) return (NhanVien) resp.getData();
+            return null;
+        }
         return nhanVienDAO.findById(maNV);
     }
 
@@ -40,6 +55,11 @@ public class NhanVienService {
      * Thêm nhân viên mới
      */
     public boolean themNhanVien(NhanVien nv) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.INSERT_NHANVIEN, nv));
+            if (resp.isSuccess() && resp.getData() != null) return (Boolean) resp.getData();
+            return false;
+        }
         return nhanVienDAO.insert(nv);
     }
 
@@ -47,6 +67,11 @@ public class NhanVienService {
      * Cập nhật thông tin nhân viên
      */
     public boolean capNhatNhanVien(NhanVien nv) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.UPDATE_NHANVIEN, nv));
+            if (resp.isSuccess() && resp.getData() != null) return (Boolean) resp.getData();
+            return false;
+        }
         return nhanVienDAO.update(nv);
     }
 
@@ -54,6 +79,11 @@ public class NhanVienService {
      * Xóa nhân viên
      */
     public boolean xoaNhanVien(String maNV) {
+        if (NetworkConfig.isClientMode()) {
+            AppResponse resp = AppClient.getInstance().sendRequest(new AppRequest(RequestType.DELETE_NHANVIEN, maNV));
+            if (resp.isSuccess() && resp.getData() != null) return (Boolean) resp.getData();
+            return false;
+        }
         return nhanVienDAO.delete(maNV);
     }
